requirements.txt
numpy==1.22.4
PySide6~=6.7.2
librosa~=0.10.2.post1
soundfile~=0.12.1
pandas~=2.2.2
Levenshtein~=0.25.1
transformers~=4.43.3
pillow~=10.3.0
turkishnlp~=0.0.61